# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "summary": "Authenticate user and generate JWT token",
                "operationId": "dd51a484b7fa0f97c7680fdbde14cf5f",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "summary": "Register a new user",
                "operationId": "c7bb21a0d87049e161b8e4c9b0d909f2",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/user/create": {
            "post": {
                "summary": "create a new user",
                "operationId": "1854111c61265030a5708391e2a17273",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "summary": "Log out user and revoke JWT token",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "201": {
                        "description": "Logout successful"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/role/create": {
            "post": {
                "summary": "Create new role.",
                "operationId": "1a0b479e6348485c9d4f435b3fbf4345",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Role name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "permissions",
                        "in": "query",
                        "description": "Array of permissions to assign to the new role",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Role created successfully"
                    },
                    "403": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/roles": {
            "get": {
                "summary": "Get all roles.",
                "operationId": "e666e3bfa734dc630ef312dee5267f25",
                "responses": {
                    "201": {
                        "description": "Role created successfully"
                    },
                    "403": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/role/update/{role_id}": {
            "put": {
                "summary": "Update an existing role.",
                "operationId": "d340c64206b01f6f23e23ea68c38e580",
                "parameters": [
                    {
                        "name": "role_id",
                        "in": "path",
                        "description": "ID of the role to be updated",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "New role name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "permissions",
                        "in": "query",
                        "description": "Array of the new permissions",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role updated successfully"
                    },
                    "403": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/role/delete": {
            "delete": {
                "summary": "Delete role.",
                "operationId": "64af4376f85f45b022bb002fe4c07065",
                "parameters": [
                    {
                        "name": "role_id",
                        "in": "query",
                        "description": "Role ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Role deleted successfully"
                    },
                    "403": {
                        "description": "Cannot delete SuperAdmin role."
                    }
                }
            }
        },
        "/api/@me": {
            "get": {
                "summary": "Get logged-in user details",
                "operationId": "b36ac7ec7c620238343e79426bb0f753",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "summary": "Get all users",
                "operationId": "32f813ea7a22e2b8b862ef93084426b9",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/delete/{user_id}": {
            "delete": {
                "summary": "Delete a user",
                "operationId": "441469730e28bcb06ac4d5152118c4c5",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "The ID of the user you want to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "403": {
                        "description": "SuperAdmin user cannot be deleted"
                    }
                }
            }
        },
        "/api/user/update/{user_id}": {
            "put": {
                "summary": "Update a user",
                "operationId": "8b03adf6e1ed8e07c7bc74e0d3c9e702",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "The ID of the user you want to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The new name of the user ( Don't fill if you don't want to updat the name)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The new email of the user ( Don't fill the email if you don't want to update the email)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "The new role of the user ( Don't fill this if you don't want to update the role )",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "403": {
                        "description": "SuperAdmin user cannot be updated"
                    },
                    "422": {
                        "description": "Role does not exist"
                    }
                }
            }
        },
        "/api/user/password/new": {
            "put": {
                "summary": "Update user password",
                "operationId": "2a9dc0374c211b4c64125314c5b92631",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "The ID of the user you want to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The new password ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Password updated successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/@me/password/new": {
            "put": {
                "summary": "Update my password",
                "operationId": "6a810e29d82c56a188bb33783a8df824",
                "parameters": [
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The new password ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Password updated successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}